move to straight threadsx
