+++ packages/viewer/package.json
@@ -82,7 +82,6 @@
     "@vertexvis/scene-view-protos": "^0.8.5",
     "@vertexvis/stream-api": "1.1.0",
     "@vertexvis/utils": "1.1.0",
-    "@vertexvis/web-workers": "^0.2.1",
     "camel-case": "^4.1.2",
     "classnames": "^2.5.1",
     "date-fns": "^2.30.0",
@@ -96,7 +95,7 @@
     "regl": "^2.1.1",
     "regl-shape": "^1.1.0",
     "resize-observer": "^1.0.4",
-    "threads": "^1.7.0",
+    "threadsx": "^2.1.0",
     "tslib": "^2.8.1"
   },
   "devDependencies": {
@@ -110,7 +109,6 @@
     "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^8.5.9",
     "@vertexvis/eslint-config-vertexvis-typescript": "^0.6.2",
-    "@vertexvis/rollup-plugin-web-workers": "^0.2.1",
     "@vertexvis/typescript-config-vertexvis": "1.3.2",
     "@vertexwebsdk/build": "1.1.0",
     "abortcontroller-polyfill": "^1.7.8",
