+++ packages/viewer/src/workers/__mocks__/worker-url.ts
@@ -1,0 +1,4 @@
+export const pngDecoderWorkerUrl = new URL(
+  'png-decoder.worker.js',
+  'http://localhost',
+);
