+++ packages/viewer/src/workers/png-decoder.worker.ts
@@ -1,6 +1,6 @@
-import { defineWorker } from '@vertexvis/web-workers';
 import type { DecodedPng } from 'fast-png';
 import { decode } from 'fast-png';
+import { expose } from 'threadsx/worker';
 
 export type DecodePngBytes = ArrayBufferLike | Uint8Array<ArrayBufferLike>;
 
@@ -10,4 +10,4 @@ async function decodePng(bytes: DecodePngBytes): Promise<DecodedPng> {
   return decode(bytes instanceof Uint8Array ? bytes : new Uint8Array(bytes));
 }
 
-defineWorker(decodePng);
+expose(decodePng);
