+++ packages/viewer/src/workers/worker-url.ts
@@ -1,0 +1,4 @@
+export const pngDecoderWorkerUrl = new URL(
+  './png-decoder.worker.ts',
+  import.meta.url,
+);
