--- packages/viewer/tsconfig.docs.json
@@ -1,5 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/tsconfig",
-  "extends": "./tsconfig.json",
-  "exclude": ["dist", "node_modules", "**/*.spec.ts"]
-}
