--- packages/viewer/types/web-workers.d.ts
@@ -1,1 +1,0 @@
-import '@vertexvis/web-workers/dist/extensions';
